UI optimized for all device

+++ together/Event.swift
@@ -16,7 +16,7 @@ class Event {
     var contrebuted: Int
     var photo: UIImage
     var category: String
-    var ownerId: Int
+    var ownerId: UInt64
     var likes: Array<Int>
     var location: String
     var startTime: Date
@@ -25,7 +25,7 @@ class Event {
     
     
     
-    init(title: String = "", description: String = "", id: Int = 0, photo: UIImage = #imageLiteral(resourceName: "EventPhoto"), contrebuted: Int = 0, category: String = "", ownerId: Int = 0, likes: Array<Int> = [0], location: String = "", startTime: Date = Date(), endTime: Date = Date(), signedUsers: Array<Int> = [0]) {
+    init(title: String = "", description: String = "", id: Int = 0, photo: UIImage = #imageLiteral(resourceName: "EventPhoto"), contrebuted: Int = 0, category: String = "", ownerId: UInt64 = 0, likes: Array<Int> = [0], location: String = "", startTime: Date = Date(), endTime: Date = Date(), signedUsers: Array<Int> = [0]) {
         self.title = title
         self.id = id
         self.description = description
